random forest: correct initialization of vector 'var_type'
var_type is a vector that describes which attribute/feature is numeric,
categorical, etc…
Its size is (1 x #features+1). The first #features elements contain the
information about the features/attributes, the last element contains the
information about the label.
In our case all the features/attributes are numeric, while the label is
categorical.